Restyle project card to use flexboxes and adapt to the relocated CTAs
Use flex so that the content can grow to take up additional space.

This allows the image to be at the top of the card and the CTAs to be at the bottom instead of directly underneath the amount of space that the content takes up, resulting in a consistent look.